# openapi 2.yml
openapi: '3.0.2'
info:
  title: blog system
  description: this is an indie gaming blog system made by Yanhe Wang
  version: '1.0'
servers:
  - url: https://localhost:8080
paths:
  /login:
    get:
      tags:
        - users
      description: 'return login page'
      operationId: login
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/LoginHtml"
  /home:  
    get:
      tags:
        - users
      description: 'return home page'
      operationId: backHome
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/HomeHtml"
    post:
      tags:
        - users
      description: 'get and check user info input then return home page'
      operationId: login
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                username:
                  type: "string"
                password:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/Login"
  /register:  
    get:
      tags:
        - users
      description: 'return register page'
      operationId: getRegisterView
      requestBody:
        required: true
        content:
          text/html: 
            schema:
              type: "string"        
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/RegisterHtml"
    post:
      tags:
        - users
      description: 'get user info input and put them into database'
      operationId: register
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                username:
                  type: "string"
                password:
                  type: "string"
                rpassword:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/Register"
 
  /publish:
    post:
      tags: 
        - users
      description: 'get blog info input and put them into database'
      operationId: publish
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                blogTitle:
                  type: "string"
                blogContent:
                  type: "string"
                name:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/Publish"

  /editor:
    get:
      tags:
        - users
      description: 'return editor page'
      operationId: newBlog
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/EditorHtml"

  /reader:
    get:
      tags:
        - users
      description: 'return reader page'
      operationId: readBlog
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                blogId:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/ReaderHtml"

  /edit:
    post:
      tags: 
        - users
      description: 'edit the existing blog'
      operationId: editBlog
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                blogTitle:
                  type: "string"
                blogContent:
                  type: "string"
                name:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/EditBlog"

  /delete:
    post:
      tags: 
        - users
      description: 'delete blog info in database'
      operationId: deleteBlog
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                blogId:
                  type: "string"
                name:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                $ref: "#/components/schemas/DeleteBlog"

components:
  schemas:
    LoginHtml:
      type: "string"
    Login:
      type: "object"
      properties:
        username:
          type: "string"
        view:
          type: "string"
    RegisterHtml:
      type: "string"
    Register:
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
        view:
          type: "string"
    Publish:
      type: "object"
      properties:
        blogTitle:
          type: "string"
        blogContent:
          type: "string"
        name:
          type: "string"
        view: 
          type: "string"
    HomeHtml:
      type: "string"
    EditorHtml:
      type: "string"
    ReaderHtml:
      type: "string"
    DeleteBlog:
      type: "object"
      properties:
        view:
          type: "string"
    EditBlog:
      type: "object"
      properties:
        blogTitle:
          type: "string"
        blogContent:
          type: "string"
        name:
          type: "string"
        view:
          type: "string"